refactor: store refresh token in redis instead of mongodb
