+++ src/services/job.service.js
@@ -71,10 +71,6 @@ export class JobService {
       throw new Error(MSG.USER.NOT_FOUND);
     }
 
-    // if (!user.refreshToken) {
-    //   throw new Error(MSG.USER.NOT_LOGGED_IN);
-    // }
-
     const company = await this.companyDao.isActive(companyId);
     const isOwner = await this.companyDao.isOwner(companyId, userId);
 
@@ -121,10 +117,6 @@ export class JobService {
     if (!user) {
       throw new Error(MSG.USER.NOT_FOUND);
     }
-
-    // if (!user.refreshToken) {
-    //   throw new Error(MSG.USER.NOT_LOGGED_IN);
-    // }
 
     const company = await this.companyDao.isActive(companyId);
     const canManage = await this.companyDao.canManage(companyId, userId);
